Resolve CVE-2018-3750
Vulnerable versions: < 0.5.1
Patched version: 0.5.1

The utilities function in all versions <= 0.5.0 of the deep-extend node 
module can be tricked into modifying the prototype of Object when the 
attacker can control part of the structure passed to this function. This 
can let an attacker add or modify existing properties that will exist on 
all objects.